homogénéisation coding style

+++ src/scene_display.cpp
@@ -2,73 +2,78 @@
 #include <GL/glut.h>  
 #include <math.h>
 
+
 void disk_bis(float x, float y, float r)
 {
-    glBegin( GL_TRIANGLE_FAN );
-        glVertex2f(x, y);
-        for( int n = 0; n <= 20; ++n ) {
-            float const t = 2*M_PI*(float)n/(float)20;
-            glVertex2f(x + sin(t)*r, y + cos(t)*r);
-        }
-    glEnd();
+	glBegin(GL_TRIANGLE_FAN);
+		glVertex2f(x, y);
+		for(int n = 0; n <= 20; ++n)
+		{
+			float const t = 2.0 * M_PI * (float)n / 20.0;
+			glVertex2f(x + sin(t)*r, y + cos(t)*r);
+		}
+	glEnd();
 }
+
 
 void Scene::display()
 {
-    //Scene initialization
-    glClearColor(0.0f, 0.0f, 0.0f, 1.0f);
-    glClear(GL_COLOR_BUFFER_BIT);
+	// Scene initialization
+	glClearColor(0.0f, 0.0f, 0.0f, 1.0f);
+	glClear(GL_COLOR_BUFFER_BIT);
 
-    //Preliminary scaling : t
-    // frame [xmin,xmax]*[ymin,ymax]
-    float x_min = map.xMin;
-    float x_max = map.xMax;
-    float y_min = map.yMin;
-    float y_max = map.yMax;
-    
-    glLoadIdentity();
-    glTranslatef((x_max+x_min)/(x_min-x_max),(y_max+y_min)/(y_min-y_max),0.0);    
-    glScalef(2/(x_max-x_min),2/(y_max-y_min),0.0);
+	// Preliminary scaling : t
+	// frame [xmin,xmax]*[ymin,ymax]
+	float x_min = map.xMin;
+	float x_max = map.xMax;
+	float y_min = map.yMin;
+	float y_max = map.yMax;
+	
+	glLoadIdentity();
+	glTranslatef((x_max+x_min)/(x_min-x_max), (y_max+y_min)/(y_min-y_max), 0.0);
+	glScalef(2.0/(x_max-x_min), 2.0/(y_max-y_min), 0.0);
    
-    //TODO : delete
-    //Test
-    glColor3f(1.0,0.0,0.0);
-    disk_bis(0,0,1);
-    disk_bis(-10,0,1);
-    disk_bis(0,5,1);
+	//TODO : delete
+	// Test
+	glColor3f(1.0, 0.0, 0.0);
+	disk_bis(0, 0, 1);
+	disk_bis(-10, 0, 1);
+	disk_bis(0, 5, 1);
 
-    //Obstacle rendering
+	//Obstacle rendering
 
-    
-    //Groups (Agents) rendering
-    std::vector<Group>::iterator it;
-    for (it = groups.begin() ; it != groups.end(); ++it) {
-        (*it).draw();
-    }
+	
+	//Groups (Agents) rendering
+	std::vector<Group>::iterator it;
+	for (it = groups.begin(); it != groups.end(); ++it)
+	{
+		(*it).draw();
+	}
 
-    glFlush();
-
+	glFlush();
 }
 
-//Trick to handle the fact that display belongs to a class (scene).
-//http://stackoverflow.com/questions/5008266/callback-function-in-freeglut-from-object
+
+// Trick to handle the fact that display belongs to a class (scene).
+// http://stackoverflow.com/questions/5008266/callback-function-in-freeglut-from-object
 Scene * ptr_global_instance = NULL;
 extern "C"
-void display_callback(){
-    ptr_global_instance->display();
+void display_callback()
+{
+	ptr_global_instance->display();
 }
+
 
 //Initializes the window for the display
 void Scene::init_window(){
-     int height = 720;
-     float ratio_f = (map.yMax - map.yMin)/(map.xMax - map.xMin);
-     int ratio_i = int(ratio_f + 0.5);
-     glutInitWindowSize(height*ratio_i,height);
-     glutInitWindowPosition(0,0);
-     glutCreateWindow("Crowd Simulator V0");
-     ::ptr_global_instance = this;
-     ::glutDisplayFunc(::display_callback);
-     glutMainLoop(); 
+	 int height = 720;
+	 float ratio_f = (map.yMax - map.yMin)/(map.xMax - map.xMin);
+	 int ratio_i = int(ratio_f + 0.5);
+	 glutInitWindowSize(height*ratio_i,height);
+	 glutInitWindowPosition(0,0);
+	 glutCreateWindow("Crowd Simulator V0");
+	 ::ptr_global_instance = this;
+	 ::glutDisplayFunc(::display_callback);
+	 glutMainLoop(); 
 }
-
